requirements.txt
python-dotenv==1.0.0
discord.py==2.4.0
requests==2.31.0
transmission-rpc==7.0.8

